chore: add network logging
